+++ src/reducers/reducerFactories/reportOptionsFactory.js
@@ -1,7 +1,7 @@
 ﻿export default function(itemRoot, actionTypes, defaultState = {}) {
     return (state = defaultState, action) => {
         switch (action.type) {
-            case actionTypes[`REQUEST_${itemRoot}_REPORT`]:
+            case actionTypes[`REQUEST_${itemRoot}`]:
                 return {
                     ...action.payload.options
                 };